network: allow max bandwith above 2047 Mbps
fix #17

+++ src/NetIcon/SettingsForm.cs
@@ -38,7 +38,7 @@ namespace NetIcon
             settings.backgroundColor = Utils.ColorFromString(networkBackgroundText.Text);
             settings.borderColor = Utils.ColorFromString(networkBorderText.Text);
             settings.updateInterval = (int)networkInterval.Value;
-            settings.maxBandwithBitPerSecond = (int)(networkMaxBandwidth.Value * 1024 * 1024);
+            settings.maxBandwithBitPerSecond = (long)(networkMaxBandwidth.Value * 1024 * 1024);
 
             Close();
         }
